+++ core/src/main/java/ua/edu/chdtu/deanoffice/service/StudentDegreeService.java
@@ -2,32 +2,18 @@ package ua.edu.chdtu.deanoffice.service;
 
 import org.springframework.stereotype.Service;
 import ua.edu.chdtu.deanoffice.entity.StudentDegree;
-import ua.edu.chdtu.deanoffice.entity.StudentExpel;
-import ua.edu.chdtu.deanoffice.repository.CurrentYearRepository;
 import ua.edu.chdtu.deanoffice.repository.StudentDegreeRepository;
-import ua.edu.chdtu.deanoffice.repository.StudentExpelRepository;
 
-import java.util.Date;
 import java.util.List;
 
-import static ua.edu.chdtu.deanoffice.Constants.EXPELLED_STUDENTS_YEARS_FOR_INITIAL_VIEW;
 import static ua.edu.chdtu.deanoffice.Constants.FACULTY_ID;
-import static ua.edu.chdtu.deanoffice.Constants.SUCCESS_REASON_IDS;
 
 @Service
 public class StudentDegreeService {
     private final StudentDegreeRepository studentDegreeRepository;
-    private final StudentExpelRepository studentExpelRepository;
-    private final CurrentYearRepository currentYearRepository;
 
-    public StudentDegreeService(
-            StudentDegreeRepository studentDegreeRepository,
-            StudentExpelRepository studentExpelRepository,
-            CurrentYearRepository currentYearRepository
-    ) {
+    public StudentDegreeService(StudentDegreeRepository studentDegreeRepository) {
         this.studentDegreeRepository = studentDegreeRepository;
-        this.studentExpelRepository = studentExpelRepository;
-        this.currentYearRepository = currentYearRepository;
     }
 
     public StudentDegree getById(Integer id) {
@@ -49,14 +35,5 @@ public class StudentDegreeService {
     public StudentDegree getFirstStudentDegree(Integer studentId) {
         List<StudentDegree> studentDegreeList = this.studentDegreeRepository.findByStudentId(studentId);
         return (studentDegreeList.isEmpty()) ? null : studentDegreeList.get(0);
-    }
-
-    public List<StudentExpel> getAllExpelStudents(Integer facultyId) {
-        return this.studentExpelRepository.findAllFired(SUCCESS_REASON_IDS, getLimitYear(), facultyId);
-    }
-
-    private Date getLimitYear() {
-        int currentYear = currentYearRepository.getOne(1).getCurrYear();
-        return new Date((currentYear - EXPELLED_STUDENTS_YEARS_FOR_INITIAL_VIEW) + "/01/01");
     }
 }
