+++ core/src/main/java/ua/edu/chdtu/deanoffice/service/StudentExpelService.java
@@ -2,26 +2,35 @@ package ua.edu.chdtu.deanoffice.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import ua.edu.chdtu.deanoffice.entity.CurrentYear;
 import ua.edu.chdtu.deanoffice.entity.StudentDegree;
 import ua.edu.chdtu.deanoffice.entity.StudentExpel;
+import ua.edu.chdtu.deanoffice.repository.CurrentYearRepository;
 import ua.edu.chdtu.deanoffice.repository.StudentDegreeRepository;
 import ua.edu.chdtu.deanoffice.repository.StudentExpelRepository;
 
+import java.util.Date;
 import java.util.List;
 import java.util.stream.Collectors;
+
+import static ua.edu.chdtu.deanoffice.Constants.EXPELLED_STUDENTS_YEARS_FOR_INITIAL_VIEW;
+import static ua.edu.chdtu.deanoffice.Constants.SUCCESS_REASON_IDS;
 
 @Service
 public class StudentExpelService {
     private final StudentDegreeRepository studentDegreeRepository;
     private final StudentExpelRepository studentExpelRepository;
+    private final CurrentYearRepository currentYearRepository;
 
     @Autowired
     public StudentExpelService(
             StudentDegreeRepository studentDegreeRepository,
-            StudentExpelRepository studentExpelRepository
+            StudentExpelRepository studentExpelRepository,
+            CurrentYearRepository currentYearRepository
     ) {
         this.studentDegreeRepository = studentDegreeRepository;
         this.studentExpelRepository = studentExpelRepository;
+        this.currentYearRepository = currentYearRepository;
     }
 
     public List<StudentExpel> expelStudents(List<StudentExpel> studentExpels) {
@@ -34,5 +43,14 @@ public class StudentExpelService {
         studentDegreeRepository.save(studentDegrees);
 
         return studentExpelRepository.save(studentExpels);
+    }
+
+    public List<StudentExpel> getAllExpelStudents(Integer facultyId) {
+        return this.studentExpelRepository.findAllFired(SUCCESS_REASON_IDS, getLimitYear(), facultyId);
+    }
+
+    private Date getLimitYear() {
+        int currentYear = currentYearRepository.getOne(1).getCurrYear();
+        return new Date((currentYear - EXPELLED_STUDENTS_YEARS_FOR_INITIAL_VIEW) + "/01/01");
     }
 }
